# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ManualMovementsManager API",
...
    "termsOfService": "https://yourcompany.com/terms",
    "contact": {
      "name": "Development Team",
      "url": "https://yourcompany.com",
      "email": "dev@yourcompany.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0.0"
  },
  "paths": {
...
    "/api/v1/Product/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Obter produto por ID",
        "description": "Retorna os dados de um produto específico baseado no ID fornecido.",
        "operationId": "GetProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Define o idioma preferido (ex: en-US, pt-BR, es)",
            "schema": {
              "type": "string",
              "default": "pt-BR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado e retornado com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "ID inválido",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductByKeyResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Atualizar produto",
        "description": "Atualiza os dados de um produto existente. Valida se o código do produto é único (exceto para o próprio produto).",
        "operationId": "UpdateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Define o idioma preferido (ex: en-US, pt-BR, es)",
            "schema": {
              "type": "string",
              "default": "pt-BR"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProductRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProductRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflito - código do produto já existe",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Entidade não processável",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeProductResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Remover produto",
        "description": "Remove um produto específico baseado no ID fornecido. A remoção é lógica (soft delete).",
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Define o idioma preferido (ex: en-US, pt-BR, es)",
            "schema": {
              "type": "string",
              "default": "pt-BR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto removido com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              }
            }
          },
          "204": {
            "description": "Produto não encontrado (já removido)"
          },
          "400": {
            "description": "ID inválido",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveProductResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeProductRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChangeProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Product"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "$ref": "#/components/schemas/Exception"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataStatus": {
        "enum": [
          "created",
          "active",
          "inactive",
          "deleted"
        ],
        "type": "string"
      },
...
      "Exception": {
        "type": "object",
        "properties": {
          "targetSite": {
            "$ref": "#/components/schemas/MethodBase"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "data": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true,
            "readOnly": true
          },
          "innerException": {
            "$ref": "#/components/schemas/Exception"
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "hResult": {
            "type": "integer",
            "format": "int32"
          },
          "stackTrace": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "GetProductByKeyResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Product"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "$ref": "#/components/schemas/Exception"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ManualMovement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "launchNumber": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "cosifCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "movementDate": {
            "type": "string",
            "format": "date-time"
          },
          "userCode": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "productCosif": {
            "$ref": "#/components/schemas/ProductCosif"
          }
        },
        "additionalProperties": false
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "productCosifs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCosif"
            },
            "nullable": true
          },
          "manualMovements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualMovement"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductCosif": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "productCode": {
            "type": "string",
            "nullable": true
          },
          "cosifCode": {
            "type": "string",
            "nullable": true
          },
          "classificationCode": {
            "type": "string",
            "nullable": true
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "manualMovements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualMovement"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RemoveProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "boolean"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "$ref": "#/components/schemas/Exception"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}